pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "mysql-connector-python>=9.3.0",
    "openai-agents>=0.0.10",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.1",
]
